requirements.txt
asttokens==2.4.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
fonttools==4.53.0
fsspec==2024.6.0
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
torch==2.3.1
torchvision==0.18.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
